+++ src/main/java/com/codingchosun/backend/controller/user/SignUpController.java
@@ -1,4 +1,4 @@
-package com.codingchosun.backend.controller;
+package com.codingchosun.backend.controller.user;
 
 import com.codingchosun.backend.dto.request.RegisterUserRequest;
 import com.codingchosun.backend.dto.response.ApiResponse;
